Find a Historical Climate Data station

Search for stations in the Historical Climate Data inventory by name.

find_station(name = NULL, ignore.case = TRUE, glob = FALSE, ...)

Arguments

name

character; character string or a regular expression to be matched against known station names. See grep for details.

ignore.case

logical; by default the search for station names is not case-sensitive.

glob

logical; use wildcards (e.g. "yell*") in the name. See link{glob2rx} for details.

...

Additional arguments passed to grep.

Value

An object of class "hcd_station_list", which is a "tbl_df", containing details of any matching HCD stations.

Author

Gavin L. Simpson

Examples

find_station("Regina")
#> # A tibble: 30 × 5
#>    Name                       Province     StationID LatitudeDD LongitudeDD
#>    <fct>                      <fct>        <fct>          <dbl>       <dbl>
#>  1 REGINA                     Saskatchewan 7450            NA           NA 
#>  2 REGINA                     Saskatchewan 7451            50.4       -105.
#>  3 REGINA INT'L A             Saskatchewan 3002            50.4       -105.
#>  4 REGINA                     Saskatchewan 50877           50.4       -105.
#>  5 REGINA CO-OP UPGRADER SNOW Saskatchewan 46587           50.5       -105.
#>  6 REGINA SIAST SNOW          Saskatchewan 46588           50.4       -105.
#>  7 REGINA RCMP SNOW           Saskatchewan 46607           50.4       -105.
#>  8 REGINA INTL A              Saskatchewan 51441           50.4       -105.
#>  9 REGINA AP STN A            Saskatchewan 7452            50.4       -105.
#> 10 REGINA AP STN B            Saskatchewan 7453            50.4       -105.
#> # ℹ 20 more rows

find_station("Yellowknife")
#> # A tibble: 7 × 5
#>   Name                  Province              StationID LatitudeDD LongitudeDD
#>   <fct>                 <fct>                 <fct>          <dbl>       <dbl>
#> 1 YELLOWKNIFE A         Northwest Territories 1706            62.5       -114.
#> 2 YELLOWKNIFE A         Northwest Territories 51058           62.5       -114.
#> 3 YELLOWKNIFE OVERLAP   Northwest Territories 53898           62.5       -114.
#> 4 YELLOWKNIFE-HENDERSON Northwest Territories 45467           62.4       -114.
#> 5 YELLOWKNIFE CON MINE  Northwest Territories 8949            62.5       -114.
#> 6 YELLOWKNIFE CS        Northwest Territories 27338           62.5       -114.
#> 7 YELLOWKNIFE HYDRO     Northwest Territories 1707            62.7       -114.

find_station("Yell*", glob = TRUE)
#> # A tibble: 9 × 5
#>   Name                  Province              StationID LatitudeDD LongitudeDD
#>   <fct>                 <fct>                 <fct>          <dbl>       <dbl>
#> 1 YELLOWKNIFE A         Northwest Territories 1706            62.5       -114.
#> 2 YELLOWKNIFE A         Northwest Territories 51058           62.5       -114.
#> 3 YELLOWKNIFE OVERLAP   Northwest Territories 53898           62.5       -114.
#> 4 YELLOWKNIFE-HENDERSON Northwest Territories 45467           62.4       -114.
#> 5 YELLOWKNIFE CON MINE  Northwest Territories 8949            62.5       -114.
#> 6 YELLOWKNIFE CS        Northwest Territories 27338           62.5       -114.
#> 7 YELLOWKNIFE HYDRO     Northwest Territories 1707            62.7       -114.
#> 8 YELLOWHEAD LO         Alberta               2641            53.2       -117.
#> 9 YELLOW GRASS          Saskatchewan          3060            49.8       -104.