Read Historical Climate Data files

Reads data from CSV files download from the Canada Historical Climate Data website. The function knows about hourly, daily, and monthly data formatted files.

read_hcd(file, flags = FALSE, clean = TRUE, ...)

Arguments

file

Either a path to a file, a connection. See argument file of read_csv.

flags

Logical; should variable flag columns be preserved. The default results in these columns being dropped from the returned data.

clean

Logical; should variable names be cleaned? The default replaces the variable identifiers from the CSV files with more useful names for use in R code.

...

Further arguments passed to read_csv.

Value

A tbl_df.

Author

Gavin L. Simpson

## FIXME: if hadley fixes the bug in count_fields, add it back here as import from readr & remove utils import

Examples

## read a monthly data file
read_hcd(system.file("extdata/2855-monthly-data.csv", package = "canadaHCD"))
#> # A tibble: 515 × 16
#>    Station   StationID Longitude Latitude Date      MaxTemp MinTemp MeanTemp
#>    <chr>     <chr>         <dbl>    <dbl> <yearmon>   <dbl>   <dbl>    <dbl>
#>  1 BROADVIEW 4010879       -103.     50.4 Jan 1965     NA      NA       NA  
#>  2 BROADVIEW 4010879       -103.     50.4 Feb 1965     NA      NA       NA  
#>  3 BROADVIEW 4010879       -103.     50.4 Mar 1965     -8.1   -20.4    -14.3
#>  4 BROADVIEW 4010879       -103.     50.4 Apr 1965      7.3    -1.9      2.7
#>  5 BROADVIEW 4010879       -103.     50.4 May 1965     16       1.9      9  
#>  6 BROADVIEW 4010879       -103.     50.4 Jun 1965     21.7     9.4     15.6
#>  7 BROADVIEW 4010879       -103.     50.4 Jul 1965     24.7    10.4     17.6
#>  8 BROADVIEW 4010879       -103.     50.4 Aug 1965     24.6     9.1     16.9
#>  9 BROADVIEW 4010879       -103.     50.4 Sep 1965     10.4     0.1      5.3
#> 10 BROADVIEW 4010879       -103.     50.4 Oct 1965     14.1    -0.6      6.8
#> # ℹ 505 more rows
#> # ℹ 8 more variables: ExtremeHigh <dbl>, ExtremeLow <dbl>, TotalRain <dbl>,
#> #   TotalSnow <dbl>, TotalPrecip <dbl>, LastSnowGrnd <int>, MaxGustDir <int>,
#> #   MaxGustSpeed <chr>

## read a daily data file
read_hcd(system.file("extdata/2855-daily-data-2015.csv", package = "canadaHCD"))
#> # A tibble: 365 × 16
#>    Station   StationID Longitude Latitude Date       MaxTemp MinTemp MeanTemp
#>    <chr>     <chr>         <dbl>    <dbl> <date>       <dbl>   <dbl>    <dbl>
#>  1 BROADVIEW 4010879       -103.     50.4 2015-01-01    -4.9   -12.7     -8.8
#>  2 BROADVIEW 4010879       -103.     50.4 2015-01-02    -3.9   -19.6    -11.8
#>  3 BROADVIEW 4010879       -103.     50.4 2015-01-03   -19.5   -29.9    -24.7
#>  4 BROADVIEW 4010879       -103.     50.4 2015-01-04   -26.9   -34.9    -30.9
#>  5 BROADVIEW 4010879       -103.     50.4 2015-01-05   -19.3   -31.3    -25.3
#>  6 BROADVIEW 4010879       -103.     50.4 2015-01-06   -20.7   -25.4    -23.1
#>  7 BROADVIEW 4010879       -103.     50.4 2015-01-07   -13     -29.4    -21.2
#>  8 BROADVIEW 4010879       -103.     50.4 2015-01-08   -13.2   -21.7    -17.5
#>  9 BROADVIEW 4010879       -103.     50.4 2015-01-09   -18.9   -24.1    -21.5
#> 10 BROADVIEW 4010879       -103.     50.4 2015-01-10   -20.9   -27.5    -24.2
#> # ℹ 355 more rows
#> # ℹ 8 more variables: HeatDegDays <dbl>, CoolDegDays <dbl>, TotalRain <dbl>,
#> #   TotalSnow <dbl>, TotalPrecip <dbl>, GroundSnow <int>, MaxGustDir <int>,
#> #   MaxGustSpeed <chr>